pyproject.toml
[project]
name = "context-engineering-rag"
version = "0.0.1"
description = "This is a module implementing a rag use case with cursor."
authors = [{ name = "Theresa Fruhwuerth", email = "theresa.fruhwuerth@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "llama-index>=0.11.0",
    "llama-index-vector-stores-chroma>=0.3.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-readers-file>=0.2.0",
    "llama-index-retrievers-bm25>=0.3.0",
    "chromadb>=0.5.0",
    "arxiv>=2.0",
    "pypdf>=5.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://trummelbummel.github.io/context-engineering-rag/"
Repository = "https://github.com/trummelbummel/context-engineering-rag"
Documentation = "https://trummelbummel.github.io/context-engineering-rag/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "httpx>=0.27.0",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",
    
    "mypy>=1.19.1",
    
    "ruff>=0.14.10",
    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/context_engineering_rag"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


